pyproject.toml
[tool.poetry]
name = "distance-to-sea"
version = "0.1.0"
description = "Distance calculation from population-weighted centroid of areas in England and Wales to the nearest point on the coastline."
authors = ["Harriet Sands"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^0.14.1"
pandas = "^2.1.4"
geopy = "^2.4.1"
shapely = "^2.0.2"
pathlib = "^1.0.1"
bng-latlon = "^1.1"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest-black = "^0.3.12"
pre-commit = "^3.6.0"
pytest = "^7.4.3"
ipykernel = "^6.28.0"
toml = "^0.10.2"

[tool.isort]
profile = "black"

[tool.pylint]
max-line-length = 88

[tool.black]
max-line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
